requirements.txt
# FastAPI
fastapi~=0.115.12
uvicorn~=0.34.1

# Database and ORM
pydantic~=2.11.3
tortoise-orm~=0.25.0
aerich~=0.8.2
tomlkit~=0.13.2
aiosqlite~=0.21.0

# Telegram
python-telegram-bot~=22.0
python-telegram-bot[job-queue]~=22.0

# Environment vars
dotenv~=0.9.9
python-dotenv~=1.1.0

#yaml
pyyaml~=6.0.2

#LLM
google-genai~=1.13.0